Fixed gameVairables values in Python
Former-commit-id: e94046dd228f249cb56224f3bb23fa8ed9091054

+++ include/ViZDoomDefines.h
@@ -35,7 +35,7 @@ namespace vizdoom{
 
     struct GameState {
         unsigned int number;
-        std::vector<double> gameVariables;
+        std::vector<int> gameVariables;
 
         uint8_t * screenBuffer;
         uint8_t * depthBuffer;
